+++ src/module.js
@@ -17,6 +17,9 @@
 
     .provider('networking', function () {
       var service = {
+        setBaseUrl: setBaseUrl,
+        getBaseUrl: getBaseUrl,
+        clearBaseUrl: clearBaseUrl,
         applyBaseUrl: applyBaseUrl
       };
       var provider = {